A meidas de mejorar las Politicas aplicando el role moderator

+++ app/Http/Controllers/Pages/PageGamesController.php
@@ -4,10 +4,14 @@ namespace App\Http\Controllers\Pages;
 
 use App\Http\Controllers\Controller;
 use App\Models\Game;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Storage;
 
 class PageGamesController extends Controller
 {
+
+    use AuthorizesRequests;
+
     public function index()
     {
 
@@ -22,16 +26,23 @@ class PageGamesController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Game::class);
+
         return view('games.create');
     }
 
     public function edit(Game $game)
     {
+
+        $this->authorize('update', $game);
+
         return view('games.edit', compact('game'));
     }
 
     public function destroy(Game $game)
     {
+
+        $this->authorize('delete', $game);
 
         if ($game->image_url) {
             Storage::disk('public')->delete($game->image_url);
